Add files via upload
Man, hago commit de lo ultimo que pude hacer. Estoy muriendo del sueño.. mañana nos vemos y terminamos. Y la broma corre, solo que no saco bien las cuentas...

+++ Vertice.java
@@ -41,7 +41,7 @@ public class Vertice
  * Parametros de salida:
  * @throws altura: tipo Double, altura del vertice
 */ 
-  public double getAltura() {
+  public int getAltura() {
   	return altura;
   }
